+++ src/config/vars.ts
@@ -4,7 +4,8 @@ const path = require('path');
 // import .env variables
 require('dotenv-safe').load({
   path: path.join(__dirname, '../../.env'),
-  sample: path.join(__dirname, '../../.env.example')
+  sample: path.join(__dirname, '../../.env.example'),
+  allowEmptyValues: true
 });
 
 module.exports = {
@@ -13,7 +14,9 @@ module.exports = {
   socketEnabled: ['1', 'true', 'yes'].indexOf(process.env.SOCKET_ENABLED || '') >= 0,
   JWT_SECRET: process.env.JWT_SECRET,
   JWT_EXPIRATION_MINUTES: process.env.JWT_EXPIRATION_MINUTES,
+  UPLOAD_LIMIT: 5, // MB
   SLACK_WEBHOOK_URL: process.env.SLACK_WEBHOOK_URL,
+  EMAIL_TEMPLATE_BASE: './src/templates/emails/',
   EMAIL_FROM_SUPPORT: process.env.EMAIL_FROM_SUPPORT,
   EMAIL_MAILGUN_API_KEY: process.env.EMAIL_MAILGUN_API_KEY,
   EMAIL_MAILGUN_DOMAIN: process.env.EMAIL_MAILGUN_DOMAIN,